pyproject.toml
[project]
name = "gestionconference"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref>=3.11.0",
    "django>=6.0",
    "djangorestframework>=3.15.0",
    "djangorestframework-simplejwt>=5.4.0",
    "fastmcp>=2.13.3",
    "pillow>=11.0.0",
]
